+++ src/main/java/electrodynamics/compatibility/jei/utils/label/types/TimeLabelWrapperConstant.java
@@ -18,7 +18,7 @@ public class TimeLabelWrapperConstant extends AbstractLabelWrapper {
 	@Override
 	public Component getComponent(AbstractRecipeCategory<?> category, Object recipe) {
 
-		return ChatFormatter.getChatDisplayShort((double) ticks / 20.0, DisplayUnit.TIME_SECONDS);
+		return ChatFormatter.getChatDisplayShort(ticks / 20.0, DisplayUnit.TIME_SECONDS);
 
 	}
